requirements.txt
llama_index
python-dotenv
openai
PyPDF2
langchain
transformers